requirements.txt
pillow==9.1.0
django-formtools==2.3
django-crispy-forms==1.14.0